Implement 'Contains' Equality
